Bug 860457 - use uintptr_t instead of unsigned long in celldata.h; r=dholbert
UltraBlame original commit: efaa46a0e3e69a10fe55038e69dd33bca93f55b9

+++ layout/tables/celldata.h
@@ -8,6 +8,7 @@
 #include "nsISupports.h"
 #include "nsCoord.h"
 #include "mozilla/gfx/Types.h"
+#include "mozilla/StandardInteger.h"
 
 class nsTableCellFrame;
 class nsCellMap;
@@ -136,10 +137,9 @@ protected:
   
   
   
-  
   union {
     nsTableCellFrame* mOrigCell;
-    unsigned long     mBits;
+    uintptr_t         mBits;
   };
 };
